requirements.txt
streamlit 
pandas
matplotlib
seaborn
scapy
speedtest-cli
plotly